Fixes + cleanup
