+++ src/icons/Gem.tsx
